fix(fsck): respect shallow boundaries for reachability
Shallow repos list boundary commits in .git/shallow; parents beyond that
edge are intentionally absent. fsck must not walk those parents or it
reports missing objects and exits 2, which breaks t5542-push-http-shallow
when the hybrid lib-httpd wrapper runs fsck through grit.

- Add grit_lib::shallow::load_shallow_boundaries and reuse from rev_list
- Skip parent traversal at shallow boundary commits in fsck walk
