+++ data/test-files.csv
@@ -963,7 +963,7 @@ t5540-http-push-webdav	t5	skip	0	0	0	false	ok	2
 t5541-http-push-smart	t5	skip	20	0	0	false	timeout	0
 t5541-remote-subcommands	t5	yes	5	4	1	false	ok	0
 t5542-push-advanced	t5	yes	4	4	0	true	ok	0
-t5542-push-http-shallow	t5	yes	3	1	2	false	ok	0
+t5542-push-http-shallow	t5	yes	3	3	0	true	ok	0
 t5543-atomic-push	t5	yes	13	7	6	false	ok	0
 t5544-pack-objects-hook	t5	yes	7	3	4	false	ok	0
 t5545-push-options	t5	yes	13	5	8	false	ok	0
