+++ grit-lib/src/lib.rs
@@ -78,6 +78,7 @@ pub mod repo;
 pub mod rerere;
 pub mod rev_list;
 pub mod rev_parse;
+pub mod shallow;
 #[cfg(unix)]
 pub mod simple_ipc;
 pub mod sparse_checkout;
