+++ logs/2026-04-09_t5542-shallow-fsck.md
@@ -1,0 +1,21 @@
+# t5542-push-http-shallow — shallow `git fsck` via hybrid wrapper
+
+## Symptom
+
+`t5542-push-http-shallow` failed test 2 (`push to shallow repo via http`). The body runs `git fsck` on the bare shallow remote; with `lib-httpd.sh`, non-HTTP `git` invocations go to grit.
+
+## Root cause
+
+`grit fsck` walked full commit parents from refs and treated missing parents beyond `.git/shallow` boundaries as `missing tree` errors, exiting with code 2 (`ExplicitExit`). That broke the `git fsck &&` chain in the test (POSIX `set -e` treats exit 2 as failure).
+
+## Fix
+
+- Added `grit_lib::shallow::load_shallow_boundaries` (reads `.git/shallow`).
+- `walk_reachable` in `fsck` skips enqueueing parents for commits listed as shallow boundaries (same idea as `rev_list`).
+- Moved shallow loading out of `rev_list` into the shared module.
+
+## Verification
+
+- `./t5542-push-http-shallow.sh` — 3/3 pass
+- `./scripts/run-tests.sh t5542-push-http-shallow.sh` — updates CSV/dashboards
+- `cargo test -p grit-lib --lib` — pass
